split store to make it scalable
