+++ src/App.tsx
@@ -7,13 +7,13 @@ import MainNavigation from './components/Navigation/MainNavigation';
 import './app.scss';
 import SideDrawer from './components/Navigation/SideDrawer';
 import Backdrop from './components/Backdrop/Backdrop';
-import { useStateValue } from './Store';
+import { useStateValue } from './Store/Store';
 
 function App() {
-    const { state } = useStateValue();
+    const { uiState } = useStateValue().state;
 
     let backdrop;
-    if (state.isSideDrawOpen) {
+    if (uiState.isSideDrawOpen) {
         backdrop = <Backdrop />;
     }
     return (
