+++ src/components/Navigation/SideDrawer.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import NavigationItems from './NavigationItems';
 
 import './sideDrawer.scss';
-import { useStateValue } from '../../Store';
+import { useStateValue } from '../../Store/Store';
 
 const SideDrawer = () => {
-    const { state } = useStateValue();
+    const { uiState } = useStateValue().state;
 
     let sideDrawerClasses = ['side-drawer'];
-    if (state.isSideDrawOpen) {
+    if (uiState.isSideDrawOpen) {
         sideDrawerClasses.push('side-drawer__open');
     }
     return (
