+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
 import App from './App';
-import { StateProvider } from './Store';
+import { StateProvider } from './Store/Store';
 
 ReactDOM.render(
     <StateProvider>
